refac pennylane algs
